requirements.txt
bitsandbytes==0.43.1
datasets==2.19.1
hydra-core==1.3.2
matplotlib==3.8.4
numpy==1.24.3
omegaconf==2.3.0
pandas==2.2.2
tokenizers==0.19.1
torch==2.3.1
transformers==4.41.2
